requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
sqlparse==0.4.2
whitenoise==6.2.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
django-filter==23.1
djangorestframework==3.14.0
drf-spectacular==0.26.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.3
MarkupSafe==2.1.2
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.28.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
psycopg2==2.8.6